# Introduce `FuncInfoRef`, a wrapped raw pointer to UserFunctionInfo.