--- a/src/builtin.rs
+++ b/src/builtin.rs
@@ -125,7 +125,7 @@ pub fn require(vm: &mut VM, args: &[Value], _this: Value) -> VMValueResult {
     vm.script_info.push((id, script_info));
 
     vm.prepare_context_for_function_invokation(
-        &module_info,
+        module_info,
         Some(vm.global_environment),
         args: &[Value],
         Value::undefined(),