--- a/src/gc.rs
+++ b/src/gc.rs
@@ -1,7 +1,7 @@
 use crate::vm::{
-    constant, exec_context,
-    exec_context::{EnvironmentRecord, ExecContext, LexicalEnvironment},
+    constant,
     jsvalue::{function, object, prototype, value::Value},
+    vm::*,
 };
 use rustc_hash::{FxHashMap, FxHashSet};
 use std::hash::{Hash, Hasher};
@@ -34,8 +34,7 @@ impl Hash for GcTargetKey {
 pub struct MemoryAllocator {
     allocated_memory: MarkMap,
     pub allocated_size: usize,
-    pub total_allocated_size: usize,
-    pub total_collected_size: usize,
+    pub collected_size: usize,
     pub roots: MarkSet,
     locked: MarkSet,
     pub state: GCState,
@@ -74,8 +73,7 @@ impl MemoryAllocator {
         MemoryAllocator {
             allocated_memory: MarkMap::default(),
             allocated_size: 0,
-            total_allocated_size: 0,
-            total_collected_size: 0,
+            collected_size: 0,
             roots: MarkSet::default(),
             locked: MarkSet::default(),
             state: GCState::Initial,
@@ -88,7 +86,6 @@ impl MemoryAllocator {
         let data_size = mem::size_of_val(&data);
         let ptr = Box::into_raw(Box::new(data));
         self.allocated_size += data_size;
-        self.total_allocated_size += data_size;
         self.allocated_memory.insert(GcTargetKey(ptr), self.white);
         ptr
     }
@@ -97,25 +94,22 @@ impl MemoryAllocator {
 impl MemoryAllocator {
     pub fn mark(
         &mut self,
-        global: exec_context::LexicalEnvironmentRef,
+        global: LexicalEnvironmentRef,
         object_prototypes: &prototype::ObjectPrototypes,
         constant_table: &constant::ConstantTable,
-        //stack: &Vec<BoxedValue>,
-        cur_context: &exec_context::ExecContext,
-        saved_context: &Vec<exec_context::ExecContext>,
+        cur_context: &ExecContext,
+        saved_context: &Vec<ExecContext>,
     ) {
         self.counter += 1;
-        if self.counter < 1 {
+        if self.counter < 100 {
             return;
         };
         self.counter = 0;
+
         let mut markset = MarkSet::default();
 
         self.state = match self.state {
             GCState::Initial => {
-                // println!("initial");
-                // println!("before {:?}", self.allocated_memory);
-
                 markset.insert(GcTargetKey(global.as_ptr()));
                 global.initial_trace(&mut markset);
 
@@ -128,18 +122,11 @@ impl MemoryAllocator {
                 object_prototypes.array.initial_trace(&mut markset);
 
                 constant_table.initial_trace(&mut markset);
-                /*
-                                for val_boxed in stack {
-                                    let val: Value = (*val_boxed).into();
-                                    val.initial_trace(&mut markset);
-                                }
-                */
+
                 for context in saved_context {
                     context.initial_trace(&mut markset);
                     context.this.initial_trace(&mut markset);
                 }
-
-                // println!("initial mark: {:?}", markset);
 
                 self.white = self.white.flip_white();
 
@@ -148,27 +135,21 @@ impl MemoryAllocator {
                 GCState::Marking
             }
             GCState::Marking => {
-                // println!("start marking: {:?}", markset);
                 let roots = std::mem::replace(&mut self.roots, FxHashSet::default());
                 for root in roots {
                     self.allocated_memory.insert(root, MarkState::Black);
                     unsafe { &*root.0 }.trace(self, &mut markset);
                 }
 
-                // println!("marking: {:?}", markset);
-
                 self.roots = markset;
 
                 if self.roots.len() == 0 {
-                    // println!("all marked: {:?}", self.allocated_memory);
                     GCState::ReadyToSweep
                 } else {
                     GCState::Marking
                 }
             }
             GCState::ReadyToSweep => {
-                // println!("before {:?}", self.allocated_memory.len());
-
                 let white = self.white;
                 let mut size = 0;
                 self.allocated_memory.retain(|obj, mark| {
@@ -182,9 +163,8 @@ impl MemoryAllocator {
                     size += unsafe { Box::from_raw(obj.0).free() };
                     false
                 });
-                self.total_collected_size += size;
+                self.collected_size += size;
                 self.allocated_size -= size;
-                // println!("now allocated objects: {:?}", self.allocated_memory.len());
 
                 GCState::Initial
             }
@@ -247,8 +227,8 @@ macro_rules! mark_if_white {
     }};
 }
 
-impl GcTarget for ExecContext {
-    fn initial_trace(&self, markset: &mut MarkSet) {
+impl ExecContext {
+    pub fn initial_trace(&self, markset: &mut MarkSet) {
         mark!(markset, self.lexical_environment.as_ptr());
         mark!(markset, self.variable_environment.as_ptr());
         for env in &self.saved_lexical_environment {
@@ -259,7 +239,7 @@ impl GcTarget for ExecContext {
             val.initial_trace(markset);
         }
     }
-
+    /*
     fn trace(&self, allocator: &mut MemoryAllocator, markset: &mut MarkSet) {
         mark_if_white!(allocator, markset, self.lexical_environment.as_ptr());
         mark_if_white!(allocator, markset, self.variable_environment.as_ptr());
@@ -267,15 +247,12 @@ impl GcTarget for ExecContext {
             mark_if_white!(allocator, markset, env.as_ptr());
         }
     }
-
-    fn free(&self) -> usize {
-        mem::size_of::<exec_context::ExecContext>()
-    }
+    */
 }
 
 impl GcTarget for LexicalEnvironment {
     fn initial_trace(&self, markset: &mut MarkSet) {
-        fn trace_record(record: &exec_context::EnvironmentRecord, markset: &mut MarkSet) {
+        fn trace_record(record: &EnvironmentRecord, markset: &mut MarkSet) {
             match record {
                 EnvironmentRecord::Declarative(record)
                 | EnvironmentRecord::Function { record, .. }
@@ -300,7 +277,7 @@ impl GcTarget for LexicalEnvironment {
     fn trace(&self, allocator: &mut MemoryAllocator, markset: &mut MarkSet) {
         fn trace_record(
             allocator: &mut MemoryAllocator,
-            record: &exec_context::EnvironmentRecord,
+            record: &EnvironmentRecord,
             markset: &mut MarkSet,
         ) {
             match record {
@@ -338,7 +315,7 @@ impl GcTarget for ::std::ffi::CString {
     }
 }
 
-impl GcTarget for Value {
+impl Value {
     fn initial_trace(&self, markset: &mut MarkSet) {
         match self {
             Value::Object(obj) => {
@@ -357,10 +334,6 @@ impl GcTarget for Value {
             Value::String(s) => mark_if_white!(allocator, markset, *s),
             _ => {}
         }
-    }
-
-    fn free(&self) -> usize {
-        mem::size_of::<Value>()
     }
 }
 
@@ -384,7 +357,7 @@ impl GcTarget for object::ObjectInfo {
     }
 }
 
-impl GcTarget for object::Property {
+impl object::Property {
     fn initial_trace(&self, markset: &mut MarkSet) {
         match self {
             object::Property::Data(object::DataProperty { val, .. }) => val.initial_trace(markset),
@@ -406,13 +379,9 @@ impl GcTarget for object::Property {
             }
         }
     }
-
-    fn free(&self) -> usize {
-        mem::size_of::<object::Property>()
-    }
 }
 
-impl GcTarget for object::ObjectKind {
+impl object::ObjectKind {
     fn initial_trace(&self, markset: &mut MarkSet) {
         match self {
             object::ObjectKind::Function(func_info) => match func_info.kind {
@@ -454,13 +423,9 @@ impl GcTarget for object::ObjectKind {
             object::ObjectKind::Ordinary => {}
         }
     }
-
-    fn free(&self) -> usize {
-        mem::size_of::<object::ObjectKind>()
-    }
 }
 
-impl GcTarget for constant::ConstantTable {
+impl constant::ConstantTable {
     fn initial_trace(&self, markset: &mut MarkSet) {
         for const_ in &self.table {
             match const_ {
@@ -468,13 +433,5 @@ impl GcTarget for constant::ConstantTable {
                 _ => {}
             }
         }
-    }
-
-    fn trace(&self, _allocator: &mut MemoryAllocator, _markset: &mut MarkSet) {
-        panic!()
-    }
-
-    fn free(&self) -> usize {
-        panic!()
     }
 }