This PR:
 - Implement `FuncInfoRef` as a wrapped raw pointer to `UserFunctionInfo`. This allows significant reduction of the size of `FunctionObjectKind` and `ExecContext`, and save much time to copying function objects and ExecContext, especially in invoking JS functions (approx. 50% in `CALL` instruction).
- Some refactoring includes moving test code for VM execution (previously in `main.rs`) to the test crate(./tests).

```
pub enum FunctionObjectKind {
    User {
        /// Internal slot \[\[Environment\]\]
        outer_env: Option<LexicalEnvironmentRef>,
        info: FuncInfoRef,
    },
    Builtin(BuiltinFuncTy),
}

#[derive(Clone, Debug)]
pub struct UserFunctionInfo {
    pub func_name: Option<String>,

    /// Unique id for many purposes
    pub func_id: FunctionId,

    /// Module id
    pub module_func_id: FunctionId,

    /// Internal slot \[\[FormalParameters\]\]
    pub params: Vec<FunctionParameter>,

    /// Variable declared names
    pub var_names: Vec<String>,

    /// Lexically declared names
    pub lex_names: Vec<String>,

    /// Declared functions to initialize
    pub func_decls: Vec<FuncInfoRef>,

    /// Bytecode to execute
    pub code: ByteCode,

    /// Exception table
    pub exception_table: Vec<Exception>,

    ...and so on
}

#[derive(Clone, Debug, Copy)]
pub struct FuncInfoRef(*mut UserFunctionInfo);
```

benchmark (examples/fibo.js)
```
<before>
# performance analysis
total execution time:    5553963 microsecs
gc time:                 1101062 microsecs
Inst          total %    ave.time / inst
...
Call          38.45 %       2077 nsecs
Return         8.04 %        434 nsecs
GetValue      18.66 %        336 nsecs

<after>
# performance analysis
total execution time:    4013104 microsecs
gc time:                  860049 microsecs
Inst          total %    ave.time / inst
...
Call          23.59 %        921 nsecs
Return         8.17 %        319 nsecs
GetValue      26.02 %        338 nsecs
```